move getAnvilResult from ForgeHooks to ForgeHooksClient
